fix/userInfo : modify userInfo

+++ src/pages/MyPage/UserInfo/index.js
@@ -3,8 +3,14 @@ import { useNavigate, useLocation } from "react-router-dom";
 import { useRecoilValue, atom, useRecoilState } from "recoil";
 import dayjs from "dayjs";
 import { useMount } from "react-use";
+import { validateForm } from "../../Auth/Register/utils";
 import { useMe } from "../../../hooks";
-import { updateUser, getUsersMe, deleteUser, updateMe } from "../../../api/User";
+import {
+  updateUser,
+  getUsersMe,
+  deleteUser,
+  updateMe,
+} from "../../../api/User";
 import { userState } from "../../../state";
 import { ImageProfile2 } from "../../../assets/images/profileImages";
 import { Button, Input, Toast } from "../../../components";
@@ -18,16 +24,22 @@ const UserInfo = () => {
   const [openImgModal, setOpenImgModal] = useState(false);
   const [floatToast, setFloatToast] = useState(false);
   const [toastMsg, setToastMsg] = useState("");
+  const [err, setErr] = useState({});
+  const [showPsw, setShowPsw] = useState("text");
   const [form, setForm] = useState({
-    nickname: '',
-    email: '',
-    password: '',
-    birth: '',
-    profileImage: '',
+    nickname: "",
+    email: "",
+    password: "",
+    birth: "",
+    profileImage: "",
   });
 
   const navigate = useNavigate();
-  const location = useLocation();
+
+  const onClickPsw = () => {
+    setShowPsw(!showPsw);
+  }
+
 
   const onChangeInfo = (e) => {
     const { name, value } = e.target;
@@ -40,49 +52,58 @@ const UserInfo = () => {
   //내 정보 불러오기
   const onGetMe = async () => {
     const response = await getUsersMe();
-    if(response.status === 200) {
+    if (response.status === 200) {
       setMe(response.data);
-    }
-  }
-
-  //회원탈퇴
-  const onDelete = async () => {
-    const response = await deleteUser(user?.id);
-    if (response.status === 204) {
-      alert('삭제 완료 되었습니다');
-      navigate('/');
-    } else {
-      alert('오류!');
     }
   };
 
-  useMount(()=>{
+  //회원탈퇴
+  const onDelete = async () => {
+    if (confirm("회원 탈퇴 하시겠습니까?")) {
+      const response = await deleteUser(user?.id);
+      if (response.status === 204) {
+        alert("삭제 완료 되었습니다");
+        navigate("/");
+        (location || window.location || document.location).reload();
+      } else {
+        toast("fail")
+        return;
+      }
+    }
+  };
+
+  useMount(() => {
     onGetMe();
-  })
+  });
 
   //저장
   const onSubmit = async (e) => {
     // if (e && e.preventDefault) { e.preventDefault(); }
     e.preventDefault();
-    
+
+    if (err.email || err.password) {
+      toast("errText");
+      return;
+    }
+
     const myInfoData = {
       nickname: form?.nickname,
       email: form?.email,
       password: form?.password,
       birth: form?.birth,
-      profileImage: form?.profileImage
+      profileImage: form?.profileImage,
     };
 
     try {
       const responsePatch = await updateMe(myInfoData);
       if (responsePatch.status === 204) {
         onGetMe();
-        //toast("save");
+        toast("save");
       }
     } catch (err) {
       const errData = err.response.data;
       alert(errData.message);
-      //toast("fail");
+      toast("fail");
     }
   };
 
@@ -91,6 +112,10 @@ const UserInfo = () => {
     setOpenImgModal(!openImgModal);
   };
 
+  const cancel = () => {
+    //임시
+    (location || window.location || document.location).reload();
+  };
 
   const toast = (msg) => {
     if (!floatToast) {
@@ -106,7 +131,11 @@ const UserInfo = () => {
       }, 2000);
     }
   }, [floatToast]);
-  
+
+  useEffect(() => {
+    setErr(validateForm(form));
+  }, [form]);
+
   useEffect(() => {
     setForm({
       nickname: me?.nickname,
@@ -138,34 +167,40 @@ const UserInfo = () => {
             />
             <p onClick={onClickImg}>프로필 사진 변경하기</p>
           </article>
-          <article className={styles.infoWrapper}>
+          <article id="registerForm" className={styles.infoWrapper}>
             <div>
               <Input
                 name="nickname"
                 label="닉네임"
+                placeholder={"새로운 닉네임을 입력해주세요"}
                 value={form?.nickname}
                 onChange={onChangeInfo}
                 className={styles.inputText}
               />
             </div>
             <div>
+              {err && <p className={styles.errText}>{err.email}</p>}
               <Input
                 name="email"
                 label="이메일"
+                placeholder={"새로운 이메일을 입력해주세요"}
                 value={form?.email}
+                //errorText={err.email}
                 onChange={onChangeInfo}
                 className={styles.inputText}
               />
             </div>
             <div>
+              {err && <p className={styles.errText}>{err.password}</p>}
               <Input
                 name="password"
                 label="비밀번호"
                 type="password"
+                placeholder={"새로운 비밀번호를 입력해주세요"}
                 value={form?.password}
                 onChange={onChangeInfo}
                 className={styles.inputText}
-              />
+               />
             </div>
             <div>
               <Input
@@ -185,14 +220,17 @@ const UserInfo = () => {
             <Button
               color="secondary"
               children="취소"
-              onClick={() => toast("cancel")}
+              onClick={() => {
+                cancel();
+                toast("cancel");
+              }}
             />
             <Button
               color="primary"
               children="저장"
               onClick={(e) => {
                 onSubmit(e);
-                toast("save");
+                //toast("save");
               }}
             />
             <Toast children={toastMsg} float={floatToast} />
